feat: add SidebarControls component and integrate it into the layout; refactor SidebarToggle and update styles in ChatHeader and Greeting components
