+++ app/(chat)/layout.tsx
@@ -1,6 +1,7 @@
 import { cookies } from 'next/headers';
 
 import { AppSidebar } from '@/components/sidebar/app-sidebar';
+import { SidebarControls } from '@/components/sidebar/sidebar-controls';
 import { SidebarInset, SidebarProvider } from '@/components/ui/sidebar';
 import { auth } from '../../server/auth';
 import Script from 'next/script';
@@ -23,6 +24,7 @@ export default async function Layout({
       />
       <SidebarProvider defaultOpen={!isCollapsed}>
         <AppSidebar variant="inset" user={session?.user} />
+        <SidebarControls />
         <SidebarInset className="md:peer-data-[variant=inset]:shadow-none transition-transition-[width,height] md:peer-data-[variant=inset]:rounded-b-none md:peer-data-[variant=inset]:m-0 overflow-hidden">
           {children}
         </SidebarInset>
