+++ components/chat/chat-header.tsx
@@ -1,13 +1,5 @@
 'use client';
-import { useRouter } from 'next/navigation';
-import { useWindowSize } from 'usehooks-ts';
-
-import { SidebarToggle } from '@/components/sidebar/sidebar-toggle';
-import { Button } from '@/components/ui/button';
-import { PlusIcon } from '../icons';
-import { useSidebar } from '../ui/sidebar';
 import { memo } from 'react';
-import { Tooltip, TooltipContent, TooltipTrigger } from '../ui/tooltip';
 import { type VisibilityType, VisibilitySelector } from './visibility-selector';
 
 function PureChatHeader({
@@ -21,39 +13,13 @@ function PureChatHeader({
   selectedVisibilityType: VisibilityType;
   isReadonly: boolean;
 }) {
-  const router = useRouter();
-  const { open } = useSidebar();
-
-  const { width: windowWidth } = useWindowSize();
-
   return (
-    <header className="flex sticky top-0 py-1.5 items-center px-2 md:px-2 gap-2 rounded-t-xl">
-      <SidebarToggle className="border border-input md:group-has-[[data-state=expanded]]/sidebar-wrapper:hidden" />
-
-      {(!open || windowWidth < 768) && (
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <Button
-              variant="outline"
-              className="order-2 md:order-1 md:px-2 px-2 md:h-fit ml-auto md:ml-0"
-              onClick={() => {
-                router.push('/');
-                router.refresh();
-              }}
-            >
-              <PlusIcon />
-              <span className="md:sr-only">New Chat</span>
-            </Button>
-          </TooltipTrigger>
-          <TooltipContent>New Chat</TooltipContent>
-        </Tooltip>
-      )}
-
+    <header className="flex sticky top-0 py-1.5 items-center px-2 gap-2 rounded-t-xl">
       {!isReadonly && (
         <VisibilitySelector
           chatId={chatId}
           selectedVisibilityType={selectedVisibilityType}
-          className="order-1 md:order-3 ml-auto"
+          className="ml-auto"
         />
       )}
     </header>
